+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+### master
+- run all *.tmux plugin files as executables
+
 ### v0.0.1, 2014-05-21
 - get TPM up and running